backend/requirements.txt
django==5.2
djangorestframework==3.16
django-filter==25.1
drf_yasg[validation]==1.21.10
djangorestframework-simplejwt==5.5.0
django-mail-templated==2.6.5
python-jose[cryptography]==3.5.0
drf-recaptcha
pytest==8.3.5
pytest-django==4.11.1
black
flake8